+++ src/pyfx/view/view_frame.py
@@ -21,11 +21,14 @@ class ViewFrame(PopUpLauncher):
             current_mini_buffer=default_footer,
         ))
 
-    def focus(self, widget_name):
-        self.original_widget.set_focus(widget_name)
-
     def size(self, widget_name):
         return self.original_widget.size(widget_name)
+
+    def switch(self, widget_name, focus):
+        if focus:
+            self.original_widget.set_focus(widget_name)
+        else:
+            self.original_widget.set_no_focus(widget_name)
 
     # This is a workaround we used to be able to popup autocomplete window in
     # query bar