+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Add a warning bar to show some interactive information, such as when pressing undefined key.
+
 ### Changed
 - Switched to an Emacs-alike UI frame to always show the information bar.